Add catch for unexpected input and associated test
To Do for decimal lat long was removed as this is uneccesary since
maps platforms support the degree/minute/seconds format that the
module transmits.

The parsing will be moved to a separate module so that all the serial
reader will do is read the bytes and transmit them to a remote
server. This will reduce the power consumption of the tracker.