Add HTTP server for website and associated tests
The string parsing is now handled entirely by an HTTP server to which
the tracker posts its data. This will allow the tracker to use less
power than if performing parsing on the device.

The server will serve the website (to be added), as well as providing
an API to get the latest marker position.