Add correct GPS data and HTTP connection handling
The GPS module transmits its location in DDMM.MMMM format, not the
DDMM.SSSS format as previously assumed. This commit implements the
correction to this assumption, now parsing the data to give the
correct location.

The HTTP connections are also closed in order to avoid the gps client
on the tracker from panicking when too many file descriptors are
open.