Handle closed sockets for gps client
The gps client crashes if the remote server is not running or the if
the connection is otherwise refused. This is because it attempts to
close the response body at the defer command. If there's no
connection in the first place then there is no response and the defer
statement tries to dereference a null pointer.

This change puts the close command explicitly inside an if/else
statement checking for the error.