Protect receivers when one of them starts blocking
Previously, we would just write to the channel. However, if the receiver
channel were full for any reason, then the subsequent receivers would
have their data streams blocked as well. In order to protect them from
this, we now write the data in a separate goroutine. Now, we will
time out when a single send's duration exceeds some threshold.